update travel office

+++ app/Livewire/Dashboard/PurchesInvoiceCreate.php
@@ -115,7 +115,7 @@ class PurchesInvoiceCreate extends Component
 
     public function calculateRemaining()
     {
-        $this->paid = $this->paid ?? 0; // التأكد من أن paid ليس null
+        $this->paid = (int) ($this->paid ?? 0); // التأكد من أن paid ليس null
         $this->remaining = ($this->paid > $this->total_price) ? 0 : $this->total_price - $this->paid;
     }
